requirements.txt
Django>=4.0,<5.0
psycopg2>=2.9
gunicorn>=20.0
sorl-thumbnail
Pillow>=8.0.0
whitenoise>=6.0.0